% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_query.R
\name{create_query}
\alias{create_query}
\title{Creates a Dataset Query String}
\usage{
create_query(
  search_in = NULL,
  search_for = NULL,
  order_by = NULL,
  asc = FALSE,
  select = NULL,
  date_start = NULL,
  date_end = NULL,
  date_var = NULL,
  filter_query = NULL
)
}
\arguments{
\item{search_in}{The column(s) the function should search in for the substring specified in `search_for`}

\item{search_for}{The substring the function should search for in the columns specified in `search_in`. If `search_in` is NULL then it will be searched in all columns. This parameter cannot be NULL if `search_in` is specified.}

\item{order_by}{Order the results by a column}

\item{asc}{If TRUE the results will be ordered in ascending order. If FALSE the order will be descending. This parameter is only relevant if `order_by` is specified. Default is FALSE}

\item{select}{One or multiple columns that should be selected and be part of the final result. Can be a vector or a single string.}

\item{date_start}{A string in the form of YYYY-MM-DD or YYYY that represents the start of the date range. `date_var` must be specified, otherwise this parameter won't be evaluated. If `date_end` is not specified, all dataset records from this date on will be returned.}

\item{date_end}{A string in the form of YYYY-MM-DD or YYYY that represents the start of the date range. `date_var` must be specified, otherwise this parameter won't be evaluated. If `date_start` is not specified, all dataset records up to this date will be returned.}

\item{date_var}{The variable to which `date_end` and `date_end` refer. Prameter will only be evaluated if either `date_end` or `date_start` is specified}

\item{filter_query}{A query string where binary operators can be used (see \href{https://help.opendatasoft.com/apis/ods-explore-v2/#section/ODSQL-predicates/Text-comparison-operators}{ODS cocumentation on binary operators}). This can either be a string or a vector of multiple strings. For example 'einwohner>=100' or c('einwohner>=100','gemeinde_name="Aadorf"'). Please note that if you use the equal sign for a string, the string has to be enquoted inside the already enqueted query string(e.g. 'gemeinde_name="Aadorf"' and NOT 'gemeinde_name=Aadorf')}
}
\value{
A String containing a valid query which can later be joined with the URL to the API endpoint.
}
\description{
Creates a dataset query to query the datasets from ODS. It is possible to search for specific strings, select columns, select a date range and order the results by a variable
}
\examples{
\dontrun{

create_query(search_in="bezirk",search_for="Frauenfeld",order_by="plz",asc=FALSE,select=c("bezirk","plz","ortschaft","einwohner",),date_start="2020-01-01",date_end="2022-12-31",date_var="jahr")
}


}
