% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_version.R
\docType{data}
\name{current_version}
\alias{current_version}
\title{Current ODS Explore API Version}
\format{
a string
}
\source{
current version is v2.1 (2024-07-25)
}
\description{
The current version of the ODS Explore API
}
\details{
This dataset includes several columns of example data.
}
\keyword{datasets}
