% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catalog.R
\name{get_catalog}
\alias{get_catalog}
\title{Download the data catalog}
\usage{
get_catalog(domain = NULL)
}
\arguments{
\item{domain}{domain of the portal (z.B. data.tg.ch). If non domain is passed to the function, the currently set domain or the domain from th Environment will be used.}
}
\value{
a nested data.frame with metadata of all datasets publicly available on the platform
}
\description{
Download the catalog with all available metadata of all publicly available datasets from the given domain
}
\examples{
\dontrun{
get_catalog("data.tg.ch")

# If domain is already set
get_catalog()
}

}
