% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{get_dataset}
\alias{get_dataset}
\title{Get Dataset}
\usage{
get_dataset(domain = NULL, dataset_id)
}
\arguments{
\item{domain}{domain of the portal (z.B. data.tg.ch)}

\item{dataset_id}{the dataset_id of the given dataset. The dataset_id can be taken from the respective catalog that can be accessed via `get_catalog()`}
}
\value{
A data.frame containing the requested daaset from Opendatasoft.
}
\description{
Download a dataset as data.frame by its dataset_id
}
\examples{
\dontrun{
get_dataset(domain = "data.tg.ch","sk-stat-112")
}
}
