% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_domain.R
\name{get_domain}
\alias{get_domain}
\title{Get the domain currently used for API Calls}
\usage{
get_domain()
}
\value{
domain the user set by using set_domain() or the domain the user set in the .Renviron file under ODS_API_DOMAIN
}
\description{
Get the domain currently used for API Calls
}
