% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fields.R
\name{get_fields}
\alias{get_fields}
\title{Get Dataset Fields}
\usage{
get_fields(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=get_metadata]{get_metadata}}
  \describe{
    \item{\code{dataset_id}}{the dataset_id of the given dataset. The dataset_id can be taken from the respective catalog that can be accessed via `get_catalog()`}
    \item{\code{domain}}{domain of the portal (z.B. data.tg.ch)}
  }}
}
\value{
A list of fields from the metadata of the given dataset_id
}
\description{
Get Dataset Fields
}
\examples{
\dontrun{
get_fields("sk-stat-111")
}
}
