% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Get Metadata of Dataset}
\usage{
get_metadata(domain = NULL, dataset_id)
}
\arguments{
\item{domain}{domain of the portal (z.B. data.tg.ch)}

\item{dataset_id}{the dataset_id of the given dataset. The dataset_id can be taken from the respective catalog that can be accessed via `get_catalog()`}
}
\value{
A [list()] of metadata of the given dataset_id, including fields of the dataset
}
\description{
Get Metadata of Dataset
}
\examples{
\dontrun{
get_metadata("sk-stat-111")
}
}
