% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metas.R
\name{get_metas}
\alias{get_metas}
\title{Get Dataset Metas}
\usage{
get_metas(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=get_metadata]{get_metadata}}
  \describe{
    \item{\code{dataset_id}}{the dataset_id of the given dataset. The dataset_id can be taken from the respective catalog that can be accessed via `get_catalog()`}
    \item{\code{domain}}{domain of the portal (z.B. data.tg.ch)}
  }}
}
\value{
A list of metadata (without the fields) from the given dataset_id. Every List element represents on of the different metadata types on Opendatasoft.
}
\description{
Get Dataset Metas
}
\examples{
\dontrun{
get_metas("sk-stat-111")
}
}
