% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_ods_domain.R
\name{is_ods_domain}
\alias{is_ods_domain}
\title{Checks if the given domain is an ODS domain}
\usage{
is_ods_domain(domain)
}
\arguments{
\item{domain}{a domain (for example data.tg.ch)}
}
\value{
cleaned domain if the domain is an ODS Domain (status code == 200) and NULL otherwise.
}
\description{
Performs an API Call and checks whether the resulting status code is 200
}
\examples{
\dontrun{
is_ods_domain("data.tg.ch")
#TRUE
}
}
