% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odsAPI-package.R
\docType{package}
\name{odsAPI-package}
\alias{odsAPI}
\alias{odsAPI-package}
\title{odsAPI: An API wrapper package to download data directly from data portals created by Opendatasoft <https://www.opendatasoft.com/en/>}
\description{
Use the API of the data platforms created by Opendatasoft. Download Open Government Data (OGD), the data catalog, specific records or perform queries on datasets.
}
\keyword{internal}
