% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_catalog.R
\name{query_catalog}
\alias{query_catalog}
\title{Query Catalog with ODSQL}
\usage{
query_catalog(domain = NULL, query)
}
\arguments{
\item{domain}{domain of the portal (z.B. data.tg.ch)}

\item{query}{a query based on the ODSQL Syntax}
}
\value{
A Dataframe with the content from the defined query
}
\description{
Instead of downlaoding the whole catalog and all variables, it is also possible to use ODSQL and perform specific queries.
}
\examples{
\dontrun{
query_catalog(domain = "data.tg.ch", query = "select=dataset_id,title&where=suggest(dataset_id,'sk')&limit=100")

}
}
