% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_catalog.R
\name{search_catalog}
\alias{search_catalog}
\title{Search the Catalog for specific Searchterms}
\usage{
search_catalog(
  domain = NULL,
  searchterm,
  search_in = "title",
  select_cols = NULL
)
}
\arguments{
\item{domain}{domain of the portal (z.B. data.tg.ch)}

\item{searchterm}{A string you want to search for in the catalog}

\item{search_in}{Columns in which the function should search for the given searchterm. This can either be one or several columns. The columns must be present in the select_cols param, too. The default column to search in is the dataset title (`title`)}

\item{select_cols}{The columns the function should return. If none is given, the function will return all columns from the catalog.}
}
\value{
A data.frame containing the matching datasets
}
\description{
Search for specific searchterms in one or more columns of the catalog. Return only those columns yoou select in the select_cols argument.
}
\examples{
\dontrun{

set_domain("data.tg.ch")

search_catalog(searchterm = "siedlung",select_cols = c("dataset_id","title"))
}



}
