% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_domain.R
\name{set_domain}
\alias{set_domain}
\title{Set the domain for the API Calls}
\usage{
set_domain(domain)
}
\arguments{
\item{domain}{domain of the portal (z.B. data.tg.ch)}
}
\description{
Set the domain for the API Calls
}
