% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_domain.R
\name{check_domain}
\alias{check_domain}
\title{Check the domain}
\usage{
check_domain(domain)
}
\arguments{
\item{domain}{}
}
\value{
a ceaned domain without https/http/ prefix
}
\description{
Check the domain
}
\examples{

\notrun{
check_domain("data.tg.ch")
}


}
