% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dataset.R
\name{query_dataset}
\alias{query_dataset}
\title{Query dataset}
\usage{
query_dataset(
  .domain = NULL,
  dataset_id,
  group_by_fields = NULL,
  order_by_fields = NULL,
  select_fields = NULL,
  predicates = NULL,
  key = Sys.getenv("ODS_KEY"),
  ...
)
}
\arguments{
\item{.domain}{Domain of the ODS API.}

\item{dataset_id}{ID of the dataset to query.}

\item{group_by_fields}{Vector of fields to group by. See [ODSQL Group by clause](https://help.opendatasoft.com/apis/ods-explore-v2/#section/Opendatasoft-Query-Language-(ODSQL)/Group-by-clause).}

\item{order_by_fields}{Vector of fields to order by. See [ODSQL Order by clause](https://help.opendatasoft.com/apis/ods-explore-v2/#section/Opendatasoft-Query-Language-(ODSQL)/Order-by-clause).}

\item{select_fields}{Vector of fields to be selected. See [ODSQL select clause](https://help.opendatasoft.com/apis/ods-explore-v2/#section/Opendatasoft-Query-Language-(ODSQL)/Select-clause).}

\item{predicates}{See [ODSQL predicates](https://help.opendatasoft.com/apis/ods-explore-v2/#section/ODSQL-predicates).}

\item{key}{API key for authentication.}

\item{...}{Where clause parameters. See [ODSQL where clause](https://help.opendatasoft.com/apis/ods-explore-v2/#section/Opendatasoft-Query-Language-(ODSQL)/Where-clause). You can also use `from` and `to` to select a timespan. The function will automatically use the first variable of type date and apply the timespan on that variable.}
}
\value{
A dataset.
}
\description{
Query a dataset and use [ODSQL Syntax](https://help.opendatasoft.com/apis/ods-explore-v2/#section/Opendatasoft-Query-Language-(ODSQL)) to refine your query. With this function, it is possible to download only the needed part of the dataset.
}
\examples{
\dontrun{
dataset <- query_dataset(
  dataset_id = "example_dataset",
  group_by_fields = c("field1", "field2"),
  order_by_fields = c("field3"),
  select_fields = c("field1", "field2"),
  .domain = "your_domain.opendatasoft.com",
  key = "your_api_key",
  from = "2020-01-01",
  to = "2021-01-01"
)
print(dataset)
}
}
