% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_catalog.R
\name{search_catalog}
\alias{search_catalog}
\title{Search the Catalog for specific Searchterms}
\usage{
search_catalog(domain = NULL, search_for, search_in = "title", select = NULL)
}
\arguments{
\item{domain}{domain of the portal (z.B. data.tg.ch)}

\item{search_for}{The substring the function should search for in the columns specified in `search_in`. If `search_in` is NULL then it will be searched in all columns. This parameter cannot be NULL if `search_in` is specified.}

\item{search_in}{The column(s) the function should search in for the substring specified in `search_for`. The default column to search in is the dataset title (`title`)}

\item{select}{One or multiple columns that should be selected and be part of the final result. Can be a vector or a single string.}
}
\value{
A data.frame containing the matching datasets
}
\description{
This is a wrapper for query_catalog: search for specific searchterms in one or more columns of the catalog. Return only those columns you select in the select_cols argument.
}
\examples{
\dontrun{

set_domain("data.tg.ch")

search_catalog(search_for = "siedlung",select = c("dataset_id","title"))
}



}
